CRAYSAT-1403: Add initial support for BOS v2 API
This includes changes to the `BOSClient` API client. The `BOSClient`
class has been renamed to `BOSClientCommon`, and the functionality of
the `BOSClient` class has been replaced with a new `BOSV1Client`, a
subclass of `BOSClientCommon`. Another subclass was added for
`BOSV2Client`  in preparation to add functionality exclusive to the BOS
V2 API.

Command line and configuration file options were added to specify which
version of the BOS API to use.

Test Description:
Ran unit tests.